Fixes #35388 - Drop unnecessary timestamps only if column exists
Fix for the issue when :remove_column tries to remove
column that does not exists.

+++ db/migrate/20220515130414_drop_unnecessary_timestamps.rb
@@ -1,6 +1,21 @@
 class DropUnnecessaryTimestamps < ActiveRecord::Migration[6.0]
-  def change
-    remove_column :taxable_taxonomies, :created_at, :datetime
-    remove_column :taxable_taxonomies, :updated_at, :datetime
+  def up
+    if ActiveRecord::Base.connection.column_exists?(:taxable_taxonomies, :created_at)
+      remove_column :taxable_taxonomies, :created_at, :datetime
+    end
+
+    if ActiveRecord::Base.connection.column_exists?(:taxable_taxonomies, :updated_at)
+      remove_column :taxable_taxonomies, :updated_at, :datetime
+    end
+  end
+
+  def down
+    unless ActiveRecord::Base.connection.column_exists?(:taxable_taxonomies, :created_at)
+      add_column :taxable_taxonomies, :created_at, :datetime
+    end
+
+    unless ActiveRecord::Base.connection.column_exists?(:taxable_taxonomies, :updated_at)
+      add_column :taxable_taxonomies, :updated_at, :datetime
+    end
   end
 end
